@@ -84,7 +84,7 @@ namespace DatingApp.Api.Controllers
             {
                 var photoToReturn = _mapper.Map<PhotoForReturnDto>(photo);
 
-                return CreatedAtRoute("GetPhoto", new { id = photo.Id }, photoToReturn);
+                return CreatedAtRoute("GetPhoto", new {userId, id = photo.Id }, photoToReturn);
             }
             return BadRequest("Could not add the photo");
 
