@@ -5,11 +5,12 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Hosting;
 
 namespace DattingApp.Api {
     public class Program {
         public static void Main (string[] args) {
-            var host = CreateWebHostBuilder (args).Build ();
+            var host = CreateHostBuilder (args).Build ();
             using (var scope = host.Services.CreateScope ()) {
                 var services = scope.ServiceProvider;
 
@@ -18,17 +19,19 @@ namespace DattingApp.Api {
                     context.Database.Migrate ();
                     Seed.SeedUsers (context);
 
-                    host.Run();
                 } catch (Exception ex) {
                     var logger = services.GetRequiredService<ILogger<Program>> ();
                     logger.LogError (ex, "An error occured during migration");
 
                 }
             }
+            host.Run ();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder (string[] args) =>
-            WebHost.CreateDefaultBuilder (args)
-            .UseStartup<Startup> ();
+        public static IHostBuilder CreateHostBuilder (string[] args) =>
+            Host.CreateDefaultBuilder (args)
+            .ConfigureWebHostDefaults (webBuilder => {
+                webBuilder.UseStartup<Startup> ();
+            });
     }
 }
